Django/rest_blog/config/urls.py
"""
URL configuration for config project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""

from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import TokenRefreshView, TokenObtainPairView

from blog import views
from config.schema import schema_view
from user import views as user_views

urlpatterns = [
    path("admin/", admin.site.urls),
    path("viewset-api/", include("blog.urls")),
    path("api/", include("blog.urls.view_set_urls")),
    path("generic-api/", include("blog.urls.generics_urls")),
    path("", views.BlogListView.as_view(), name="blog_list"),
    path("create/", views.BlogCreateView.as_view(), name="blog_create"),
    path("signup/", user_views.SignUpAPIView.as_view(), name="sign_up"),
    path("token", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh", TokenRefreshView.as_view(), name="token_refresh"),
    path("token/verify", TokenObtainPairView.as_view(), name="token_verify"),
    path(
        "swagger<format>/", schema_view.without_ui(cache_timeout=0), name="schema-json"
    ),
    path(
        "swagger/",
        schema_view.with_ui("swagger", cache_timeout=0),
        name="schema-swagger-ui",
    ),
    path("redoc/", schema_view.with_ui("redoc", cache_timeout=0), name="schema-redoc"),
]
